사용 안 하는 파일 삭제 (#66)
